Update delete implementation

+++ app/src/main/java/com/example/papb_team5/MainActivity.kt
@@ -1,15 +1,21 @@
 package com.example.papb_team5
 
+import android.content.DialogInterface
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.widget.PopupMenu
+import android.widget.TextView
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.papb_team5.adapter.scheduleItemAdapter
 import com.example.papb_team5.adapter.toDoItemAdapter
+import com.example.papb_team5.data_entity.Schedule
 import com.example.papb_team5.data_entity.Task
 import com.example.papb_team5.room_database.Constant
+import com.example.papb_team5.room_database.ScheduleRoomDatabase
 import com.example.papb_team5.room_database.TaskRoomDatabase
 import kotlinx.android.synthetic.main.activity_detail_tugas.*
 import kotlinx.android.synthetic.main.activity_detail_tugas.btn_elipsis_layout
@@ -24,7 +30,9 @@ import kotlinx.coroutines.withContext
 class MainActivity : AppCompatActivity() {
 
     val db by lazy { TaskRoomDatabase(this) }
+    val db2 by lazy { ScheduleRoomDatabase (this) }
     lateinit var tasksAdapter: toDoItemAdapter
+    lateinit var schedulesAdapter: scheduleItemAdapter
 
     /*
     private val newTaskActivityRequestCode = 1
@@ -36,57 +44,15 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        //popupMenu()
-
         setupListener()
         setupRecyclerView()
-
-/*
-        //val recyclerView = findViewById<RecyclerView>(R.id.todo_recycler)
-        //val adapter = toDoItemAdapter(arrayListOf())
-        //val adapter = TaskItemAdapter()
-        //recyclerView.adapter = toDoItemAdapter(this, arrayListOf())
-        //recyclerView.layoutManager = LinearLayoutManager(this)
-
-        //recyclerView?.setHasFixedSize(true)*/
-
-        /*
-        taskViewModel.allTasks.observe(this) { tasks ->
-            // Update the cached copy of the words in the adapter.
-            tasks.let { adapter.submitList(it) }
-        }*/
-
-        /*
-        bottomNavigationView.setOnNavigationItemSelectedListener {
-            when(it.itemId){
-                //R.id.home->setCurrentFragment(mainFragment)
-
-                /*
-                R.id.home-> {
-                    val intent = Intent(this@MainActivity, fragment_home::class.java)
-                    startActivity(intent)
-                }
-                */
-
-                //.id.calendar->setCurrentFragment()
-                //R.id.profile->setCurrentFragment(profileFragment)
-                R.id.profile->{
-                    val intent = Intent(this@MainActivity, profileActivity::class.java)
-                    startActivity(intent)
-                }
-            }
-            true
-        }*/
+        setupRecyclerView2()
     }
-
-    /*
-    private fun popupMenu(){
-        val popupMenu = PopupMenu(applicationContext, t)
-    }*/
 
     override fun onStart() {
         super.onStart()
         loadTask()
+        loadSchedule()
     }
 
     fun loadTask(){
@@ -99,9 +65,22 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
+    fun loadSchedule() {
+        CoroutineScope(Dispatchers.IO).launch{
+            val schedules = db2.scheduleDao().getAllSchedules()
+            Log.d("MainActivity", "dbresponse: $schedules")
+            withContext(Dispatchers.Main){
+                schedulesAdapter.setData(schedules)
+            }
+        }
+    }
+
     fun setupListener() {
         fab.setOnClickListener{
             intentEdit(0, Constant.TYPE_CREATE)
+        }
+        txt_addSchedule.setOnClickListener{
+            intentScheduleEdit(0, Constant.TYPE_CREATE)
         }
     }
 
@@ -121,6 +100,22 @@ class MainActivity : AppCompatActivity() {
         )
     }
 
+    fun intentScheduleView(Id: Int, intentType: Int){
+        startActivity(
+            Intent(applicationContext, NewScheduleActivity::class.java)
+                .putExtra("intent_id", Id)
+                .putExtra("intent_type", intentType)
+        )
+    }
+
+    fun intentScheduleEdit(Id: Int, intentType: Int){
+        startActivity(
+            Intent(applicationContext, NewScheduleActivity::class.java)
+                .putExtra("intent_id", Id)
+                .putExtra("intent_type", intentType)
+        )
+    }
+
 
 
     private fun setupRecyclerView(){
@@ -135,17 +130,70 @@ class MainActivity : AppCompatActivity() {
                 intentEdit(task.id, Constant.TYPE_UPDATE)
             }
             override fun onDelete(task: Task) {
-                CoroutineScope(Dispatchers.IO).launch{
-                    db.taskDao().deleteTask(task)
-                    loadTask()
-                }
+                deleteDialog(task)
             }
         })
         todo_recycler.apply{
             layoutManager = LinearLayoutManager(applicationContext)
             adapter = tasksAdapter
         }
-        //todo_recycler?.setHasFixedSize(true)
     }
 
+    private fun setupRecyclerView2(){
+        schedulesAdapter = scheduleItemAdapter(this, arrayListOf(), object: scheduleItemAdapter.OnAdapterListener{
+            override fun onClick(schedule: Schedule){
+                intentScheduleView(schedule.id, Constant.TYPE_READ)
+            }
+            override fun onUpdate(schedule: Schedule){
+                intentScheduleEdit(schedule.id, Constant.TYPE_UPDATE)
+            }
+            override fun onDelete(schedule: Schedule){
+                deleteScheduleDialog(schedule)
+            }
+        })
+        schedule_recycler.apply{
+            layoutManager = LinearLayoutManager(applicationContext, LinearLayoutManager.HORIZONTAL, false)
+            adapter = schedulesAdapter
+        }
+    }
+
+    private fun deleteDialog(task: Task){
+        val alertDialog = AlertDialog.Builder(this)
+
+        alertDialog.apply{
+            setTitle("Konfirmasi")
+            setMessage("Yakin hapus ${task.taskTitle}?")
+            setNegativeButton("Batal") { dialog, i ->
+                dialog.dismiss()
+            }
+            setPositiveButton("Hapus") { dialog, i ->
+                dialog.dismiss()
+                CoroutineScope(Dispatchers.IO).launch{
+                    db.taskDao().deleteTask(task)
+                    loadTask()
+                }
+            }
+        }
+        alertDialog.show()
+    }
+
+    private fun deleteScheduleDialog(schedule: Schedule){
+        val alertDialog = AlertDialog.Builder(this)
+
+        alertDialog.apply{
+            setTitle("Konfirmasi")
+            setMessage("Yakin hapus ${schedule.scheduleTitle}?")
+            setNegativeButton("Batal") { dialog, i ->
+                dialog.dismiss()
+            }
+            setPositiveButton("Hapus") { dialog, i ->
+                dialog.dismiss()
+                CoroutineScope(Dispatchers.IO).launch{
+                    db2.scheduleDao().deleteSchedule(schedule)
+                    loadSchedule()
+                }
+            }
+        }
+        alertDialog.show()
+    }
 }
